pyproject.toml
[project]
name = "cangui"
version = "0.1.0"
description = "CAN Network GUI Application"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "PySide6>=6.6.0",
    "pyqtgraph>=0.13.7",
    "python-can[serial]>=4.4.0",
    "cantools>=39.4.0",
    "odxtools>=8.0.0",
    "udsoncan>=1.25.0",
    "can-isotp>=2.0.0",
    "numpy>=1.26.0",
]
